+++ src/MixTelematicsLaravelService.php
@@ -2,17 +2,13 @@
 
 namespace Aislandener\MixTelematicsLaravel;
 
-use Aislandener\MixTelematicsLaravel\Models\Driver;
 use Aislandener\MixTelematicsLaravel\Services\ActiveEventsService;
 use Aislandener\MixTelematicsLaravel\Services\AssetService;
 use Aislandener\MixTelematicsLaravel\Services\DriversService;
 use Aislandener\MixTelematicsLaravel\Services\GroupService;
 use Aislandener\MixTelematicsLaravel\Services\PositionService;
-use Carbon\Carbon;
-use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\Http;
-use Jumbojett\OpenIDConnectClient;
-use Jumbojett\OpenIDConnectClientException;
+use Aislandener\MixTelematicsLaravel\Services\TokenService;
+use Aislandener\MixTelematicsLaravel\Services\TripService;
 
 class MixTelematicsLaravelService
 {
@@ -22,12 +18,14 @@ class MixTelematicsLaravelService
     private AssetService $assetService;
     private GroupService $groupService;
     private PositionService $positionService;
+    private TripService $tripService;
 
     public function __construct(ActiveEventsService $activeEventsService,
                                 DriversService      $driversService,
                                 AssetService        $assetService,
                                 GroupService        $groupService,
                                 PositionService     $positionService,
+                                TripService         $tripService,
     )
     {
         $this->driversService = $driversService;
@@ -35,6 +33,7 @@ class MixTelematicsLaravelService
         $this->assetService = $assetService;
         $this->groupService = $groupService;
         $this->positionService = $positionService;
+        $this->tripService = $tripService;
     }
 
     /**
@@ -69,9 +68,20 @@ class MixTelematicsLaravelService
         return $this->groupService;
     }
 
+    /**
+     * @return PositionService
+     */
     public function positions(): PositionService
     {
         return $this->positionService;
+    }
+
+    /**
+     * @return TripService
+     */
+    public function trips(): TripService
+    {
+        return $this->tripService;
     }